+++ src/components/spartan/SInputAmount/SInputAmount.stories.ts
@@ -37,6 +37,11 @@ export default {
             description: 'The list of currencies to be displayed in the selector.',
             table: { type: { summary: 'string[]' }, category: 'Props' },
         },
+        minorUnitMode: {
+            control: 'boolean',
+            description: 'Whether to display the amount in minor unit mode.',
+            table: { type: { summary: 'boolean' }, category: 'Props' },
+        },
 
         // Events
         'update:modelValue': {
@@ -58,7 +63,7 @@ export default {
             control: { type: null },
             table: { type: { summary: null }, category: 'Events' },
             description: 'The event emitted when the currency changes.',
-        }
+        },
     },
 };
 
@@ -73,6 +78,7 @@ export const Default = createDefault({
         symbol: false,
         suffixCurrency: false,
         currencies,
+        minorUnitMode: true,
     },
     setup: () => {
         const value = ref(22.99);
@@ -80,7 +86,7 @@ export const Default = createDefault({
 
         return { value, currency, currencies };
     },
-    template: `<SInputAmount :symbol="args.symbol" :suffixCurrency="args.suffixCurrency" :currencies="args.currencies" v-model:currency="currency" v-model="value" @info="console.log" />`,
+    template: `<pre>{{value}}</pre><input v-model="value" /><SInputAmount :symbol="args.symbol" :minorUnitMode="args.minorUnitMode" :suffixCurrency="args.suffixCurrency" :currencies="args.currencies" v-model:currency="currency" v-model="value" @info="console.log" />`,
     transform: (args) =>
         `<SInputAmount ${sourceBinding(args)}  v-model=value v-model:currency="currency" :currencies="currencies" />`,
 });
@@ -96,6 +102,32 @@ export const Base = createVariation({
     },
     template: `<!-- currency: USD -->
 <SInputAmount v-model="value" v-model:currency="currency"/>`,
+});
+
+export const minorUnitMode = createVariation({
+    components: { SInputAmount },
+    containerClass: 'flex flex-col gap-2',
+    setup: () => {
+        const value = ref(22.99);
+        const valueMinor = ref(22.99);
+        const currency = ref('USD');
+
+        return { value, valueMinor, currency };
+    },
+    template: `<!-- currency: USD -->
+<div class="flex items-center gap-2">
+    <div class="flex w-[200px]">
+        <SInputAmount v-model="valueMinor" v-model:currency="currency" minorUnitMode suffixCurrency />
+    </div>
+    <span>Model value (minor unit mode): {{valueMinor}}</span>
+</div>
+
+<div class="flex items-center gap-2">
+    <div class="flex w-[200px]">
+        <SInputAmount v-model="value" v-model:currency="currency" suffixCurrency />
+    </div>
+    <span>Model value: {{value}}</span>
+</div>`,
 });
 
 export const WithSuffix = createVariation({