+++ src/components/spartan/SInputAmount/types.ts
@@ -11,7 +11,7 @@ type AmountInfo = {
 
 export type TInputAmountEmits = {
     (event: 'update:currency', value: string | undefined): void;
-    (event: 'update:modelValue', value: string | undefined): void;
+    (event: 'update:modelValue', value: number | null): void;
     (event: 'info', value: AmountInfo | undefined): void;
     (event: 'change', value: number | null): void; // Required by the vue-currency-input library
 };
@@ -23,6 +23,7 @@ export type TInputAmountProps = {
     symbol?: boolean;
     suffixCurrency?: boolean;
     currencies?: Array<keyof typeof Currencies>;
+    minorUnitMode?: boolean;
 };
 
 // TODO: vite plugin prevents splitting types