+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.0.0-beta.46 - 2024-01-11
 ### Added
 - Minor unit mode now available on `SInputAmount` component.
 - `SFilter` component now supports translation for operators.